+++ CHANGELOG.md
@@ -103,6 +103,7 @@
 - PR #3294 Update to arrow-cpp and pyarrow 0.15.1
 - PR #3310 Add `row_hasher` and `element_hasher` utilities
 - PR #3286 Clean up the starter code on README
+- PR #3332 Port NVStrings replace to cudf strings column
 
 ## Bug Fixes